requirements.txt
discord.py>=2.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0